% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_loglik_methods.R
\name{adj_loglik}
\alias{adj_loglik}
\title{Log-likelihood adjustments for fitted models}
\usage{
adj_loglik(x, cluster = NULL, use_vcov = TRUE, ...)
}
\description{
This function adjusts the log-likelihood of fitted model objects based on
\href{http://doi.org/10.1093/biomet/asm015}{Chandler and Bate (2007)}.
}
