% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_loglik_methods.R
\name{anova.chantrics}
\alias{anova.chantrics}
\title{ANOVA tables: compare nested models}
\usage{
\method{anova}{chantrics}(model1, model2, ...)
}
\arguments{
\item{model1}{Object of class \code{chantrics}, as returned by
\code{\link{adj_logLik}}.}

\item{model2}{Object of class \code{chantrics}, as returned by
\code{\link{adj_logLik}}.}

\item{...}{Further objects of class \code{chantrics}, as returned by
\code{\link{adj_logLik}}, and/or parameters that will be passed to
\code{\link[chandwich]{anova.chandwich}}, and then further to
\code{\link[chandwich]{compare_models}}. The type of adjustment, out of
\code{"vertical"}, \code{"cholesky"}, \code{"spectral"}, \code{"none"}, as
specified in the parameter \code{type}, can also be specified here.}
}
\description{
\code{anova} method for \code{chantrics} objects
}
\details{
Compare two or more nested models that have been adjusted using the
\code{\link{adj_logLik}} method. It uses the adjusted likelihood ratio test
statistic (ALRTS), as described in Section 3.5 of
\href{http://doi.org/10.1093/biomet/asm015}{Chandler and Bate (2007)}.

Note that the objects specified do not have to be sorted in a specific way,
they will be sorted by the function by the number of parameters, as returned
by \code{attr(model1, "p_current")}.
}
