% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chantrics.R
\docType{package}
\name{chantrics}
\alias{chantrics}
\title{chantrics: Loglikelihood Adjustments for Econometric Models}
\description{
\code{chantrics} adjusts the loglikelihood of common econometric models for
clustered data based on the estimation process suggested in @chanbate07,
using the \href{https://cran.r-project.org/package=chandwich}{chandwich} package,
and provides convenience functions for inference on the adjusted models.
\code{adj_loglik()} adjusts the parameter covariance matrix of the models to
incorporate clustered data, and can mitigate model misspecification by
wrapping \code{chandwich::adjust_loglik} for the supported models.
}
\details{
The returned model of class \code{chantrics} can be plugged into standard model
evaluation and model comparison methods, for example \code{summary()}, \code{confint()}
and \code{anova()}, and a hypothesis test framework provided by \code{alrtest()}.

See \code{vignette("chantrics-vignette", package = "chantrics")} for an
overview of the package.
}
