% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm}
\alias{glm}
\alias{glm.nb}
\title{Loglikelihood adjustments for glm fits}
\description{
Describe glm methods here
}
\section{Supported families (within each family, any link function should work)}{

\itemize{
\item \code{gaussian}
\item \code{poisson}
\item \code{binomial}
\item \code{MASS::negative.binomial}
}

Also works for \code{\link[MASS:glm.nb]{MASS::glm.nb()}}, note that the standard errors of the theta are not adjusted.
}

\examples{
# binomial example from Applied Econometrics in R, Kleiber/Zeileis (2008)
# ==  probit  ==
data("SwissLabor", package = "AER")
swiss_probit <- glm(participation ~ . + I(age^2),
  data = SwissLabor,
  family = binomial(link = "probit")
)
summary(swiss_probit)
swiss_probit_adj <- adj_loglik(swiss_probit)
summary(swiss_probit_adj)

# == logit ==
swiss_logit <- glm(participation ~ . + I(age^2),
  data = SwissLabor,
  family = binomial(link = "logit")
)
summary(swiss_logit)
swiss_logit_adj <- adj_loglik(swiss_logit)
summary(swiss_logit_adj)
}
