% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{The \code{hello} function.}
\usage{
hello(times)
}
\arguments{
\item{times}{The number of times the phrase should be repeated.}
}
\value{
Returns nothing and prints \code{Hello, world!} to the console.
}
\description{
Function printing \code{Hello, world!} in the console.
}
\details{
Long long long long long long longlong long long long long long long long
long long long long long long long long long long
}
\section{Even more details}{

Blah blah blah blah blah blah blah blah blah blah blah blah blah blah blah blah
}

\examples{

hello(5) #prints the phrase 5 times

}
\seealso{
\code{\link[base]{rep}} combined with \code{\link[base]{cat}} for
  other repetitions of strings.
}
\concept{string repetitions}
