% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdle.R
\name{hurdle}
\alias{hurdle}
\title{Loglikelihood adjustments for pscl::hurdle fits}
\description{
Adjust the loglikelihood and the standard errors of a fitted \code{\link[pscl:hurdle]{pscl::hurdle()}} model.
}
\details{
Describe zeroinfl models here

Note that the \code{\link[pscl:hurdle]{pscl::hurdle()}} model has to be run with the option \code{x = TRUE} in order for the adjustment to execute properly
}
\section{Supported families (within each family, any link function should work)}{

\itemize{
\item \code{geometric}
\item \code{poisson}
\item \code{negbin}
\item \code{binomial} (for the zero mass distribution only)
}
}

\examples{
# hurdle model from AER, pg. 139-140
library(pscl)
data("RecreationDemand", package = "AER")
rd_hurdle <- hurdle(trips ~ . | quality + income, data = RecreationDemand,
                    dist = "negbin", x = TRUE)
summary(rd_hurdle)

# adjust model
adj_loglik(rd_hurdle)
}
