% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_loglik_methods.R
\name{logLik_vec}
\alias{logLik_vec}
\title{Evaluate loglikelihood contributions from specific observations}
\usage{
logLik_vec(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Further arguments.}
}
\description{
Generic function for calculating loglikelihood contributions from
individual observations for a fitted model.
}
