% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm}
\alias{glm}
\title{Loglikelihood adjustments for glm fits}
\description{
Describe glm methods here
}
\section{Supported families (within each family, any link function should work)}{

\itemize{
\item \code{gaussian}
\item \code{poisson}
}
}

