% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{internal}
\alias{internal}
\alias{is.error}
\alias{raise_yield_error}
\alias{get_response_from_formula}
\alias{abort_not_chantrics}
\alias{get_variable_str_from_chantrics}
\alias{get_formula_str_from_chantrics}
\alias{get_resid_df_from_chantrics}
\alias{get_additional_args_from_chantrics_call}
\title{chandwich internal functions}
\usage{
is.error(x)

raise_yield_error(model = "model", what = "property", try_this = NULL)

get_response_from_formula(x)

abort_not_chantrics(x)

get_variable_str_from_chantrics(x)

get_formula_str_from_chantrics(x)

get_resid_df_from_chantrics(x)

get_additional_args_from_chantrics_call(x)
}
\description{
Internal functions: These are not to be exported.
}
\keyword{internal}
