% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{internal}
\alias{internal}
\alias{is.error}
\alias{get_unadj_object}
\alias{raise_yield_error}
\alias{get_response_from_formula}
\alias{get_response_from_model}
\alias{get_design_matrix_from_model}
\alias{abort_not_chantrics}
\alias{get_variable_str_from_chantrics}
\alias{get_formula_str_from_chantrics}
\alias{get_additional_args_from_chantrics_call}
\alias{get_named}
\alias{get_unnamed}
\title{chandwich internal functions}
\usage{
is.error(x)

get_unadj_object(object)

raise_yield_error(model = "model", what = "property", try_this = NULL)

get_response_from_formula(x)

get_response_from_model(object)

get_design_matrix_from_model(object, type = NULL)

abort_not_chantrics(x)

get_variable_str_from_chantrics(x)

get_formula_str_from_chantrics(x)

get_additional_args_from_chantrics_call(x)

get_named(x)

get_unnamed(x)
}
\description{
Internal functions: These are not to be exported.
}
\keyword{internal}
