% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik_vec_methods.R
\name{logLik_vec}
\alias{logLik_vec}
\title{Evaluate loglikelihood contributions from specific observations}
\usage{
logLik_vec(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Further arguments.}
}
\value{
An object of class \code{"logLik_vec"} which is a numeric vector of length
\code{nobs(object)} (i.e. the number of observations in \code{object}) of the
log-likelihood of each observation. Additionally, it contains the
attributes \code{df} (model degrees of freedom) and \code{nobs} (number of
observations).
}
\description{
Generic function for calculating the log-likelihood contributions from
individual observations for a fitted model.
}
